refactor: switch to bukkit-utilities
