+++ src/main/kotlin/io/github/slaxnetwork/config/model/ConfigSpawnPointsModel.kt
@@ -1,6 +1,6 @@
 package io.github.slaxnetwork.config.model
 
-import io.github.slaxnetwork.bukkitcommon.config.model.ConfigLocationModel
+import io.github.slaxnetwork.bukkitcore.utilities.config.model.ConfigLocationModel
 import kotlinx.serialization.Serializable
 
 @Serializable
