+++ src/main/kotlin/io/github/slaxnetwork/config/model/map/BaseMapConfig.kt
@@ -1,8 +1,8 @@
 package io.github.slaxnetwork.config.model.map
 
-import io.github.slaxnetwork.bukkitcommon.config.model.ConfigBorderModel
-import io.github.slaxnetwork.bukkitcommon.config.model.ConfigLocationModel
-import io.github.slaxnetwork.bukkitcommon.config.model.map.ConfigMapMetaModel
+import io.github.slaxnetwork.bukkitcore.utilities.config.model.ConfigBorderModel
+import io.github.slaxnetwork.bukkitcore.utilities.config.model.ConfigLocationModel
+import io.github.slaxnetwork.bukkitcore.utilities.config.model.map.ConfigMapMetaModel
 import io.github.slaxnetwork.config.model.ConfigSpawnPointsModel
 import kotlinx.serialization.SerialName
