+++ src/main/kotlin/io/github/slaxnetwork/listeners/PlayerDeathListener.kt
@@ -1,6 +1,6 @@
 package io.github.slaxnetwork.listeners
 
-import io.github.slaxnetwork.bukkitcommon.config.injectConfig
+import io.github.slaxnetwork.bukkitcore.utilities.config.injectConfig
 import io.github.slaxnetwork.config.types.SoundsConfig
 import io.github.slaxnetwork.events.crown.KOTCCrownHolderDeathEvent
 import io.github.slaxnetwork.game.GameManager
