+++ src/main/kotlin/io/github/slaxnetwork/scoreboard/TestScoreboard.kt
@@ -1,11 +1,12 @@
 package io.github.slaxnetwork.scoreboard
 
-import io.github.slaxnetwork.bukkitcore.scoreboard.*
-import io.github.slaxnetwork.bukkitcore.utils.boardLine
+import io.github.slaxnetwork.bukkitcore.scoreboard.SimpleScoreboard
+import io.github.slaxnetwork.bukkitcore.utilities.scoreboard.boardComponent
+import io.github.slaxnetwork.bukkitcore.utilities.scoreboard.boardLine
 import io.github.slaxnetwork.kyouko.models.profile.Profile
 
 class TestScoreboard(private val profile: Profile) : SimpleScoreboard {
-    override val title = BoardComponent("<red>Hello, World!")
+    override val title = boardComponent { text = "<red>Hello, World!" }
 
     override val lines = listOf(
         boardLine("test", 1) {
